Refactoring progress - missing postgres eventstore

+++ src/ProjectOrigin.VerifiableEventStore/Models/EventStoreOptions.cs
@@ -5,5 +5,5 @@ namespace ProjectOrigin.VerifiableEventStore.Models;
 public class VerifiableEventStoreOptions
 {
     [Required, Range(0, 20)]
-    public long BatchSizeExponent { get; set; }
+    public int BatchSizeExponent { get; set; }
 }
